Update patch set 3
Patch Set 3:

(51 comments)

> If anything it sounds like we should delete BoolToJbool (if we agree converting a bool to a jbool explicitly is an anti-pattern)?

The point of the whole set of CLs is that you don't get *any* protection from the more than ****** C(++) semantics of implicit bool conversion if you don't also apply it to the for us "ugly" cases. That's why the CLs are up - to discuss the impact, and whether we're OK with the  bugs it leads to not looking at it.

Patch-set: 3
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>