Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> The data we have says this is not worth it over the top 10000 apps in the play store (else I would have done it this summer). You have to justify why you think the complexity tradeoff is worth it.

We had verification failures in facebook and that’s what drove this change. Also we see that there are minor improvements in facebook execution times. We will try to create a synthetic test for this and check. Are you aware of any APK that has high number of verification failures?

Patch-set: 1
Reviewer: Gerrit User 1310609 <1310609@85c56323-6fa9-3386-8a01-6480fb634889>