Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > The data we have says this is not worth it over the top 10000 apps in the play store (else I would have done it this summer). You have to justify why you think the complexity tradeoff is worth it.
> 
> We had verification failures in facebook and that’s what drove this change. Also we see that there are minor improvements in facebook execution times. We will try to create a synthetic test for this and check. Are you aware of any APK that has high number of verification failures?

From the top 10k, apps:

Overall number of class-defs:		1721523
Classes reporting soft-fails:		5594 (0.32%)
Classes with Retry or NotReady:	7103

Top 5 “biggest” cases:
Class with 587 methods and 5 failures.
Class with 313 methods and 2 failures.
Class with 264 methods and 114 failures.
Class with 226 methods and 1 failure.
Class with 224 methods and 6 failures.

In general, the failures cluster in the small range of classes with <5 methods.

At this point, it is more meaningful to punish apps producing bad code and fix the ecosystem (as most of those are not targeted to the latest release), instead of adding complexity to ART.

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>