Use smali input in 530-checker-peel-unroll
The 530-checker-peel-unroll/unrollingWhile test relies on a CFG with two
predecessors to the loop header whereas the Java source has three logical
predecessors to the loop header. Previous versions of DX and D8 would merge the
two back-edges by associating the source line of the continuation with the split
edges. Going forward D8 associates the source line of the origin with split
edges and will no longer merge the back-edges. The input program is now obtained
from the smali code to ensure no dex compiler changes will affect the CFG.

Test: art/test/testrunner/testrunner.py -t 530-checker-peel-unroll
Change-Id: If0786125fa4bac6450f92a8b1cdf6d80320f2096