Update patch set 1
Patch Set 1:

> Artem, I think the request is to move those Java methods to smali, so the R8 team doesn't get blocked again the next time they do changes around their code generation.

Yes, it would be nice that tests with hard assumptions about the DEX inputs did not use the D8 compiler, but for now it is OK for us to get this updated and more forward with the next D8 roll. I'll link this with our roll topic as soon as that CL is uploaded (currently building the next version on our archive bots).

Patch-set: 1
Reviewer: Gerrit User 1241761 <1241761@85c56323-6fa9-3386-8a01-6480fb634889>