Revert^2 "JVMTI PopFrame support"
This reverts commit f67d0cbf2743b29aa6e1be7ae2f71762b8762785.

Redefine-stress causes classes to be loaded at slightly different
places due to the deep verification needed to redefine a class. This
would cause the parts of tests 1953 and 1954 to fail due to classes
loading at unexpected times. We changed the test to skip these
test-cases when they cannot be correctly executed.

Reason for revert: Fixed issue causing tests to fail under
                   redefine-stress
Test: ./test.py --host
Test: ./art/tools/run-libjdwp-tests.sh --mode=host
Test: ./test/testrunner/testrunner.py --host --redefine-stress
Bug: 73255278
Bug: 111357976
Change-Id: I46fb4aed878bcc9398f2b73ad5040d7c97aec3ef