Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Caused by jitthreshold:0 making the compilation take place on the main thread. This means that the class-load occurs at an unexpected place (1 frame up) causing the print to occur twice.
> 
> ngeoffray@: It might be good just for consistency with regular jit to have jit-threads handle the actual compilation. Would that be an acceptable change?

Yeah I agree, sorry I overlooked this in the previous CLs (though CanCallIntoJava was already not consistent with main thread doing the compilation). Given compiling on the main thread is only used for native-debuggable (but I'm unclear about the state of its implementation), I agree it's ok to prefer consistency.

Note that we have some tests that force the compilation, and have the main thread try to do it. For those too I think we should change it too, which will allow adding a DCHECK() that only a JIT thread attempts a JIT compilation.

Patch-set: 2
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>