Don't mark new non-moving objs in young GC
During a young GC, there is a small window during which we allowed
non-moving objects on the allocation stack to be pushed to marking
stack (and thereby setting the corresponding live-bit) while
done_scanning_ is still false. This could lead to crash during hprof
heap dump in debug builds, if the dump is done before the next GC cycle.

Test: art/test/testrunner/testrunner.py --target
Test: Flash device, adb shell am dumpheap <gms|systemui|system_server>
Bug: 117426281
Change-Id: Icbcff2596241f1f87d9cebea859de3dd10041036