Update patch set 5
Patch Set 5:

For the record, with the recent CLs in now, the code should be somewhat grepable.  Say, if you search for MIRROR_OBJECT_LOCK_WORD_OFFSET, you will find the line:
  ASM_DEFINE(MIRROR_OBJECT_LOCK_WORD_OFFSET, art::mirror::Object::MonitorOffset().Int32Value())

I think that should be enough to find where the value comes from, and to add further constants if you need to.  It does not reveal the compile time value of the constant, but then again, if you have "constexpr size_t kMirrorObjectLockWordOffset = ..." in C++, you don't see the value either (for similar reason - the value depends on arch/bitness/build-config).

Patch-set: 5