Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Igor: One more thought: You mentioned x-macros. I think this is similar, but not really x-macros. We could make into proper x-macro though! (which would simultaneously avoid the need for include guards).  However, does that would require moving all defs to single file (which I think you would not like), or nested x-macros (which I would not like). So... meh.

https://en.wikibooks.org/wiki/C_Programming/Preprocessor_directives_and_macros#X-Macros

Unfortunately the name seems overloaded with this slightly different pattern:

https://en.wikipedia.org/wiki/X_Macro#Example

-----------
Being able to include other files is actually a strength because it makes the abstraction more powerful. A typical weakness of macros is being restricted to the scope they are able to be used in, but by avoiding unconditional #includes these current ASM_DEFINE macros can be used from virtually any scope.

Patch-set: 7