Update patch set 4
Patch Set 4:

Hi Neeraj, sorry for the delay reviewing - we discussed about the CL internally, and unfortunately we are not going to take it. It breaks some Java assumptions (stack traces) which we are not ready to take for a niche optimization.

I think a better place for doing such optimization is in a proguard like tool that the developer can opt into. The risk of surprise of missing stack traces should be on the developer, not the platform.

The code is nicely written though, nice work.

Patch-set: 4
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>