WIP: Split interpreter_switch_impl into smaller functions.
Split the uber switch statement into smaller functions.
Those functions are still always inlined in the switch,
so it makes little difference for the time being.

However the separation should allow us to:
 - Remove the excessive use of macros in the interpreter.
 - Add helpers to handle similar opcodes.
 - Simplify the individual opcode handlers.
 - Simplify the interactions with mterp.
 - Easily experiment (e.g. switch vs goto dispatch).

This CL includes semantic change of HANDLE_PENDING_EXCEPTION
so that it always either jumps to handler or returns to caller.
The cases where this was not already happening we likely bugs.

Test: art/test.py -b --host --interpreter --64
Change-Id: I6075b354e0439bf89061f492dd9b9cbe743145cb