Update patch set 2
Patch Set 2:

Yes, this is part of the shadow call stack feature: b/112907825

At this point we've moved past the "investigation" stage and are landing the feature, which has been approved for the next Android release.

Also, I'm not sure if it makes sense to put this change behind a flag because it is being made together with changes in other parts of the platform (see bug) which aren't (and probably can't practically be) behind a flag.

Patch-set: 2
Reviewer: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0