Update patch set 13
Patch Set 13:

I think this is ready for review. There are still TODOs/FIXMEs, but I think they are not blocking and can be addressed in follow-up CLs (especially if/when they require additional work on ART's existing build rules). For now I think it'd be good to submit this change (as soon as https://android-review.googlesource.com/c/platform/build/soong/+/800333/6 is merged), so that the ART APEX module definition is known to Soong.

Patch-set: 13
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>