Update patch set 1
Patch Set 1:

> Patch Set 1:
>a 
> (9 comments)
> 
> I'd prefer to see some tests but unfortunately we do not yet have the infrastructure for checker tests dependent on the instruction set features. That's actually one more reason why the HAndNeg/HLowestOneBit should be architecture-independent and have a non-AVX2 implementation as well.

Thank you for the review. Unfortunately these instructions are not available for x86 on the lower end processors (atom, sandybridge, ivybridge, silvermont). Due to this reason,I did not put them in nodes.h.  A non-AVX2 implementation would just be the unsimplified version (which is 2 instructions vs AVX2 which will lower to one instruction).So it would just be an overhead in combining 2 HIRS to 1 HIR in this case. So should we still go ahead and have an non AVX2 implentation as well? I have verified the functionality on host by giving "instruction-set-features=avx2" with test case in this patch   https://android-review.googlesource.com/c/platform/art/+/770421.

Patch-set: 1
Reviewer: Gerrit User 1371444 <1371444@85c56323-6fa9-3386-8a01-6480fb634889>