Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 2:
> > 
> > > This doesn't address the log tagging. This will regress the logcat back to zygote64. I don't see this addressed anywhere.
> > 
> > (See b/34867873)
> 
> I can see two things affecting the logging tag: SetDefaultTag calls and argv[0] to the program. I don't understand how this change affects either of those. Can you please explain how the regression occurs?

SetDefaultTag called in the zygote from libandroid_runtime currently has an effect on ART because they share the same libbase instance. This has the advantage of:

* Tagging affects all logging (though you can override)
* The ART abort handler will be used for all LOG(FATAL), giving whole-process diagnostics.
* Command line args to libbase logging initialization will be "shared"

> Is there a test that catches it?

As usual, there's no test anywhere. In the end, this is a Quality-of-Service-like issue, not a functional problem.

Patch-set: 3
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>