Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Logging is in a separate library, liblog, which is in the NDK. Can we move ART over to use that?
> 
> That would be nice, but we'd need to extend it with the default tag, aborter hook, etc to avoid the regressions Andreas point out.

Focussing on the buts:

The liblog variants take the tag as an argument.

ART controls when it calls abort. If other code calls abort, should we run the ART abort hook? There's not much outside of art and init that sets the hook.

Patch-set: 3
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>