Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Logging is in a separate library, liblog, which is in the NDK. Can we move ART over to use that?
> > 
> > That would be nice, but we'd need to extend it with the default tag, aborter hook, etc to avoid the regressions Andreas point out.
> 
> Focussing on the buts:
> 
> The liblog variants take the tag as an argument.

Yes, but the global state is the default tag for it, and ART seldom overrides that with an explicit tag (I see the LOG_TAG define getting set only in libart_fake and openjdkjvm).
 
> ART controls when it calls abort. If other code calls abort, should we run the ART abort hook? There's not much outside of art and init that sets the hook.

There's more stuff: the minimum logging level, and a logger hook that lets processes decide whether to send messages to stderr or logd etc.

That looks to me like good things to play ball with, for debuggability and so on, but I don't know the concrete use cases for all those features wrt ART. However if it's ok to move the libbase logging functionality (http://cs/aosp-master/system/core/base/logging.cpp) to liblog we won't need to ponder that, so I'd like to start with that question. I presume there would still need to be a slim shim in libbase to not break other users.

Patch-set: 3
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>