Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > > Patch Set 3:
> > > > > 
> > > > > > Logging is in a separate library, liblog, which is in the NDK. Can we move ART over to use that?
> > > > > 
> > > > > That would be nice, but we'd need to extend it with the default tag, aborter hook, etc to avoid the regressions Andreas point out.
> > > > 
> > > > Focussing on the buts:
> > > > 
> > > > The liblog variants take the tag as an argument.
> > > 
> > > Yes, but the global state is the default tag for it, and ART seldom overrides that with an explicit tag (I see the LOG_TAG define getting set only in libart_fake and openjdkjvm).
> > >  
> > > > ART controls when it calls abort. If other code calls abort, should we run the ART abort hook? There's not much outside of art and init that sets the hook.
> > > 
> > > There's more stuff: the minimum logging level, and a logger hook that lets processes decide whether to send messages to stderr or logd etc.
> > > 
> > > That looks to me like good things to play ball with, for debuggability and so on, but I don't know the concrete use cases for all those features wrt ART. However if it's ok to move the libbase logging functionality (http://cs/aosp-master/system/core/base/logging.cpp) to liblog we won't need to ponder that, so I'd like to start with that question. I presume there would still need to be a slim shim in libbase to not break other users.
> > 
> > The constraints are:
> > (1) the NDK exposes a C ABI for as part of stability.
> > (2) libbase has a C++ ABI which is not stable across releases.
> > 
> > It may be reasonable to just move the state related global values to libbase. We probably don't have any other conflicts than with the zygote start up logging.
> 
> That should have read:
> 
> "The constraints are:
> (1) the NDK exposes a C ABI for as part of stability.
> (2) libbase has a C++ ABI which is not stable across releases.
> 
> We might be better to fork the code into an ART namespace and be done with it." :-)

As you may have forgotten history (it probably was before your time): libbase came to be from our copy of logging macros, and we clearly *won* by giving them to system/core, as otherwise it's on us to support windows (which liblog doesn't) - do not forget that we have SDK tools to support. Moving us back will be a pain, unless it is a 1:1 up-to-date copy (a C++ jarjar ;-) ).

But copying stuff is the only way to get all the functionality back - we do not want to old crappy ALOG macros, we're pretty much bound to the LogMessage wrappers and streams everywhere. Changing that would be a major pain (and maintenance and usability regression).

Not having the abort logger will significantly screw debuggability of crashes (at least for dogfood), as our state dump (that includes actual locks) to logcat will be gone. It also means we can't do some of the testing possible right now (as there's actual an abort hook [that's normally set by AndroidRuntime]).

Patch-set: 3
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>