Replace switch interpreter macros with C++ functions.
C++ code is more maintainable than huge macros.

The "return" statement was used within the macros to skip the reminder
of the interpreted opcode.  This is slightly more tricky in functions.
All such use cases have been replaced with "return false", and the
calling functions always check the return value and propagate it up.

Test: "test.py -b -r --interpreter --host" with switch interpreter.
Change-Id: I01d861dec1c4a7e95e412fd4b662e5a4d7bd55a8