Add module `libopenjdk` to the ART APEX module.
This is so that we can remove it from `PRODUCT_PACKAGES`, i.e. no
longer install it on /system, as module `libopenjdk` depends on module
`libart`; the installation of the former would triggers the
installation of the latter, which are trying to stop (see companion
change in project platorm/build).

Test: Device boot test with ART APEX.
Bug: 113373927
Change-Id: I17a6dec4b7b4e813a6d55c4aafbad2e793bfd69d