Print the address of the object sentinel for debugging purposes.
This object is always allocated in the non-moving space, and could be
the one triggering a failing assertion in
art::gc::collector::ConcurrentCopying::ProcessMarkStackRef. Print it
as part of the information dumped when this assertions fails.

Note: This is a temporary change that should be reverted after the
corresponding bug is closed.

Test: art/test.py
Bug: 116087961
Change-Id: I678989b4ad921fefbacbfc42218e18f7302cb174