Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

Note: Your CL is fine as-is, but when relanding a CL in Android, it is customary to "revert the revert" by using Gerrit's "Revert" button (in this case, this would have meant reverting https://android-review.googlesource.com/c/platform/art/+/819174). This makes it easier to track the history of a change through reverts. (We also try to flatten CL titles like

  Revert "Revert "Revert "Revert "<Initial message>""""

into

  Revert^4 "<Initial message>"

for the sake of readability).

Patch-set: 3
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1