Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> (1 comment)
> 
> Note: Your CL is fine as-is, but when relanding a CL in Android, it is customary to "revert the revert" by using Gerrit's "Revert" button (in this case, this would have meant reverting https://android-review.googlesource.com/c/platform/art/+/819174). This makes it easier to track the history of a change through reverts. (We also try to flatten CL titles like
> 
>   Revert "Revert "Revert "Revert "<Initial message>""""
> 
> into
> 
>   Revert^4 "<Initial message>"
> 
> for the sake of readability).

Thanks for the tip. Just to try it out I did a revert through Gerrit in aosp/821441. I was a bit bummed by it adding reviewers right away though - I need to change it a bit first. You think it should be used for "non-clean" reverts like this one is?

Anyway, I'll continue with that CL and scrap this one.

Patch-set: 3
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>