Update patch set 5
Patch Set 5:

I think this should work, but the code is starting to be a little non-obvious.

It would be cleaner if we followed powers of two (e.g. call AddSamplesImpl only when we reach the next power of two).   Would that lose too much granularity on the thresholds?

Another approach would be store "countdown" in ArtMethod instead of "count".  That is, initialize it with WarmThreshold(), keep counting down to zero, re-initialize it to HotThreshold(), count down to zero again, etc...  The main trouble there is we need to store the cold/warm/hot state separately.  And the thread-specific boost weight.

Any opinions?

Patch-set: 5
Reviewer: Gerrit User 1065473 <1065473@85c56323-6fa9-3386-8a01-6480fb634889>