Update patch set 2
Patch Set 2:

Hi Jiyong, I made this crude CL to test linker namespaces for the Runtime APEX. It does what we expect - libraries are loaded completely separately in the two namespaces.

However, as you can see it gets messy because all libs are in the same directory. The separation in this example is also inaccurate because "runtime" finds all libs in its own search path which it shouldn't.

I'd like to discuss options here. Would it be feasible to split up the libs in two directories? That would make the APEX have a nonstandard directory structure, which I think the apex Soong rule currently doesn't support.

Some of the libs in the "platform" namespace is Bionic and should be used by all native users on the device. That makes this APEX a bit special, and I suspect we'll put the rules for that in /system/etc/ld.config.txt instead.

NB: This CL is a bit sloppy and loads some system libs from the APEX which shouldn't be there at all, e.g. liblog.

Patch-set: 2
CC: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>