WIP: Switch to libbase_art and put the default logd logger here.
This will allow us to plumb the liblog dependency through PALette.

This forks libbase into two separate library instances, which with the
global state in its logging code means a real observable ODR problem:
Libraries outside ART that continues to use libbase will not be affected by
the ART InitLogging call and other settings, and vice versa. The fix for
that is to move the global state to liblog, to be addressed separately.

Test: mmma art
Test: env ART_TEST_KEEP_GOING=false art/test.py --64 --optimizing
Bug: 119513342
Change-Id: I1d008655cd3b08be364dc82db6cd29ff271179b0