Update patch set 3
Patch Set 3:

(1 comment)

> > [1] https://mast.users.x20web.corp.google.com/aosp-lib-deps/2018-11-14/target-reverse-deps.html#libbase.so%20is%20a%20dependency%20of:
> 
> Thanks for persevering here and for highlighting concerns and solutions. It's all good in the grand scheme of things.
> 
> For the examples cited libnativeloader and libziparchive:
> 
> - libnativeloader is used in frameworks and art. It may only need the more minimal libbase_art? It's going to need work like libnativehelper if it's going in the APEX (C ABI).

Any library would do fine with libbase_art strictly speaking, it's just the top level binaries that need libbase for the default LogdLogger implementation. However the ODR problem becomes more severe the more libraries use it, until the binaries start using it too. Hence if we go all the way we'd basically just change all InitLogging call sites to get a LogdLogger from somewhere else, and there'd be no reason to have a separate libbase_art anymore.

If the ODR problem is solved, it doesn't matter which libbase variant is used. But the really messy bit here is that this introduces the ODR problem on host too, and there we don't have liblog to move the globals to.

> - libziparchive has some logging.h usage in the writer for errors only, but looks largely compatible with libbase_art otherwise.

The reason I pointed it out was that it's a library that sits above PALette in ART but it's used in lots of places outside ART too. If we fix the ODR it doesn't matter which lib it uses (like above), if we don't then we need to split it into libziparchive and libziparchive_art.

> The third variant here is pushing on the line being proposed. I'm in favour, but let's discuss today as you have a lot more context here.

Not sure what you are suggesting, even after the meeting. Can you please clarify?

Patch-set: 3
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>