Speed up relative_patcher_arm64_test.
This was one of the longest running tests because it had
many tests and for each test we have to create and tear down
the runtime. Refactor the macro-generated tests to use
templates and merge many individual tests. This reduces the
time to run the test from over 50s to less then 20s on host.

Also speed up relative_patcher_thumb2_test a bit.

Test: m test-art-host-gtest
Change-Id: If7ffedbca788442cebf69e1276947fde9654f4a2