Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> I don't understand enough of the NDK build to realize why you can remove the header. Are NDK headers automatically added?

we haven't actually shipped an NDK with this header yet. i only added it recently, and didn't add it to <android/api-level.h> (which is always included via <sys/cdefs.h> if you include *any* bionic header) because that was too hard. when we had a second function that ought to go there, folks complained so i'm now doing the right thing. (and, yes, it turned out to be even harder than we'd originally realized --- it's taken an entire week at this point :-/ nearly there though!)

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>