Update patch set 3
Patch Set 3:

I've been talking a bit with enh@ about this.

If there were a solution in liblog that solved this issue, that would be the ideal choice, right?  Even on the host, it seems reasonable that we implement a host version of liblog and handle all of the state there.  That would leave libbase being stateless and make it such that it doesn't matter how we link it.

Is that something we can all get behind?  I'll work to create such an interface in liblog if so.

Patch-set: 3