ART: Add boot/app image checksums.
And check app oat files against the boot image checksum
instead of checking the oat checksum. The oat checksums are
included in image checksum calculations and the primary
image checksum calculation includes the checksums of the
secondary images.

Also remove the obsolete "patch delta" to keep the size
of the ImageHeader 8-byte aligned.

Test: m test-art-host-gtest
Test: testrunner.py --host
Test: Pixel 2 XL boots.
Change-Id: If74f5f18479c44ede0503bf1911ddb9ff8f3c4f8