Update patch set 5
Patch Set 5:

Neil, my hesitation over this CL is that it gives us flexibility that may not be required. [ adding rpl@ as he is much better acquainted here, but on vacation for a few days ].

For device tests, the APEX path is fixed. For non-install location testing, it may be we're okay with chroot.

For host tests, is there something concrete that is about to break without this new path? If no, then it might be okay (if frustrating) to hold off until we find we need this.

Patch-set: 5
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>