Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> What is the reason to wait? If not, what happened?

The GC could in the process of handling weak references, and mutators shouldn't process them. There's a condition variable / locks around those to make sure the access is exclusive.

The effect of not waiting is a crash in the mutator thread, as it's not allowed to read a weak reference.

Patch-set: 7