Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)
> 
> Can you share how these profiles were generated?  What was the benefit from PGO?
> 
> The profiles should be updated as ART code in AOSP changes.  We'd also need a plan/process refreshing these profiles.



Hi, Please fine the detailed steps we followed :
=======================================
We have generated the profile files based on the android google case study published here :
https://source.android.com/devices/tech/perf/pgo#case-study-pgo-for-art

The dex2oat binary already had the support for supporting pgo profiles in its clang-build 
but art-runtime module was missing.
 
The steps explicitly we used for Intel based x86 platforms. 

=> We built the flash files with the tag "ANDROID_PGO_INSTRUMENT=art_runtime"
=> Currently we have Generated the profile files "*.profraw" for benchmarks like :
Antutu,vellamo, icy rocks & jARTband
We also profiled for launch time of some apps from Play store.
We observed 5-8% on the launch time and KPIs scores that we ran and profiled for.

But then as you see we have done this PGO generation for some KPIs only , 
so if there is any specific guideline for list of APPs that needs to be included, we can try to generate the profiles for the same with regards to the x86 platforms accordingly.

=> Using the llvm-profdata tool , convert the .profraw file to .profdata files
=> For "art-runtime" module we had only one profraw file generated. 
But for the dex2oat module , we had 3 profraw file generated with 3 unique-ids , and we merged the same with "llvm-profdata merge" tool
=======================================
Benchmark	 Profiled libraries
dex2oat	     dex2oat (executable), libart-compiler.so, libart.so
art_runtime	 libart.so
=============================

So now we have got two bechmarks based profraw files for x86 platform, and we added the same in the path : "aosp/toochain/pgo-grofiles/art" directory
These profraw files are included in another code link:
"https://android-review.googlesource.com/c/toolchain/pgo-profiles/+/836646"

=> Currently we can see there was an "arm" baseddex2oat prof file. We have now added the x86 based prof files.
As we understand from the documentation , the pgo was already enabled even for dex2oat binary and we have just added only the files for 
x86 platforms after profiling and testing only on the x86_64 targets

=> Regarding the impact on other architectures like mips, or any other architectures, the documentation says that if the profile file is present 
(currently relative to toolchain/pgo-profile path ), it will be used with PGO. 
If not available,the build just warns that this file doesn't exist by adding this file to $DIST_DIR/pgo_profile_file_missing.txt. 
=========================================================

So as of now we have done the PGO instrumentation and collected profile for some of our use case. But it would be great as you mentioned if you any specific set of Apps from AOSP guidelines and then we can generate and add the same. 
If you need the list functions/methods that have been generated in the profiles , I can send you the decompiled list of the methods.



Thanks !

Patch-set: 1
Reviewer: Gerrit User 1275029 <1275029@85c56323-6fa9-3386-8a01-6480fb634889>