Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > I don't see any reason why the JIT should be restarted in the zygote after the fork. This looks like unnecessary complexity for zero gain.
> 
> Processes could be forked and a potential JIT queue not empty. We can delete the threads when the queue is empty.

Well, that plays very well together with the request for numbers.

I don't see a reason to not finish JITing *everything* before the first fork. Otherwise you have unshared code cache troubles for a persistent process, plus overhead for having to JIT the same thing in multiple processes.

But if the important things are JITed, why re-enable the JIT? It should have touched everything. And every additional JIT just dirty-unshares more memory.

Plus, enabling it to then have to disable it on next fork takes time - and you add this time to everything, even if the JIT isn't used because AOT-ed code is available.

Patch-set: 2