Update patch set 1
Patch Set 1:

I assume eventually conscrypt will be split out to support the conscrypt module.

If we want it to eventually match the content of the runtime module then it'll include apache-xml, bouncycastle, okhttp, etc.

Is there any reason to / not to make core.art match the runtime module boot classpath entries?

Patch-set: 1