Update patch set 1
Patch Set 1:

If it's just for testing with art host then my suggestions of matching the runtime module doesn't necessarily make sense. I wasn't aware of the scope. There may be some benefit to getting them aligned eventually so "host art" is as similar to "runtime module" on device (increasing the chances of spotting behavior or performance issues). But, if they're not aligned today then it's not something to worry about here.

> > If we want it to eventually match the content of the runtime module then it'll include apache-xml, bouncycastle, okhttp, etc.
> 
> Is apache-xml needed? There is no apache-xml-testdex library at the moment.

No - apparently there's no strong requirement. Only worth doing if you want to match the runtime module on device AND there's some advantage to matching the .art image with the boot classpath on device provided by the runtime module.

> I believe core.art is for testing and therefore it should be driven by testing needs and I'd like to keep it as small as possible. I missed the libcore tests, so I thought that for testing we need just the core-{oj,libart,simple}.
> 
> OTOH, I'm trying to build infrastructure where the boot*.art can be split into logical parts and the runtime module boot classpath is the most reasonable primary boot image. It could even be shipped compiled inside the runtime module.

I have no strong opinions either way but my instinct would be to make things as similar as possible.

Patch-set: 1