X86: Use memory to do array range checks
Currently, a HBoundsCheck is fed by an HArrayLength, causing a load of
the array length, followed by a register compare.

Avoid the load when we can by comparing directly with the array length
in memory.  Implement this using an arch specific pass that replaces
the HBoundsCheck with a HX86BoundsCheckMemory that references the array
directly.  This allows us to generate code that references the length
in memory directly.

Only do this replacement when we are the only user of the ArrayLength
and it isn't visible to the environment.

Handle the special case where the array is 'null' and where an implicit
null check can't be eliminated.

This code moves the load of the length to the slow code for the failed
check, which is what we want.

Test: test.py --host --64

Change-Id: If79017cc14c838d6930b00ce0e74311f9f8ce123
Signed-off-by: Ganesh Shivaprakash <ganesh.shivaprakash@intel.com>