Add support for parsing @CorePlatformApi annotation
Creates a new category of hiddenapi flags: SpecializedApiFlags
and its first member kCorePlatformApi.

Fields/methods annotated with @CorePlatformApi are parsed by
`class2greylist` and flagged 'core-platform-api'. This flag is
recognized by `hiddenapi` and a bit is set in the hiddenapi dex
flags of that class member. ClassLinker then reads the dex flags
and sets a new runtime flag in ArtField/ArtMethod to mark the
runtime representation of the corresponding field/method.
The runtime flag will be used by hidden API logic to warn about
invalid access.

Bug: 119068555
Test: compiles, phone boots. proper tests once the flag is used
Change-Id: Ia282f704d183682683ba168a300c28179af11250